War Troll, a combat tracker and character manager for Rolemaster!
What an exciting project! We're about to build a cool web application for all
the Rolemaster (2nd edition) gamemasters out there – all three of them.

You may not know what Rolemaster is. It is an old school tabletop role-
playing game. Its 2nd edition was released in 1984 and, as far as I'm
concerned, is still the best so far.

Rolemaster's combat system is a bit different from those of more modern RPGs,
and heavily inspired (I think) by wargames. It is divided in formal phases,
during which each player must declare and resolve specific actions; this requires
to keep track of every character's activity through the entire round. It is a bit more
work – and strategic opportunities – than just sorting everybody's playing
order.

For this reason, a combat tracker could be helpful. So this is what we will
build. This will give us opportunities to play with the nice new things in the
Rails ecosystem, such as Hotwire or delegated types.

Another specifity of Rolemaster is the density of its character sheet. A Rolemaster
character is defined by _a lot_ of numbers, some of which require computations or
cross-referencing tables. Once again, having a tool to automate this would prove
useful.

So this what we're going to build: a combat tracker, and an character manager, or at
least a dynamic character sheet.

For the moment, all we have is the _shape_ of our project, in the form of a few
sketches and a use case scenario. And this is all we need! The shape gives us a
direction and enough matter to start coding and iterating. The details — in code,
design and actual functionalities — will emerge as we progress.