Introducing `Skill::Bundle` and its AR type
We'll need to store bundle of skills — values for skills represented by
the same class. For example, we need a bundle of the four "maneuver in armor"
skills, which are all represented by the `Skill::ManeuverInArmor`.

The bundle itself is nothing special — a simple hash-like structure does
the trick. Our bundle can thus be nothing more than a wrapper around a Struct.

(This means that 2 bundles of the same skills will be instances of different,
but 100% identical, classes. Should this become a problem, some caching store
— ideally thread-safe — could be implemented.)

The AR type, however, is a bit special. Since most of the attributes for a
skill do not need to be persisted — things like stat bonuses, level bonuses,
and so on are character-dependent and should be applied _after_ deserialization —, a bit of logic is needed to clean up the data before writing to the DB, and
preparing it when reading from the DB.