Tackling "primary" skills
And, doing so, encountering our first problems with the way skills and
skill bundles are managed.

As expected, some skill data is character-dependent. In particular, the
_stat_ associateds with a primary skill (`Skill::Primary.stat`) must be
passed by the `Character` object.

But this object is unavailable when the AR attribute is cast from the DB
value. Which means that the stat must be passed when _accessing the skill_
(ie. when calling a relevant method on the `Character` instance), and not
when _casting the skill from the DB_. One abstraction layer above, in a way.

This is what we already do for the stats; so it's not such a special case.
But it encourages us to reevaluate our architecture. Since _all_ the
logic for preparing the skills doesn't fit in the AR attributes layer,
should it even be there in the first place?