CT, scope 1: keeping track of phases and rounds
This commit squashes the whole work done on the first scope in the
Combat Tracker project: keeping track of the rounds, their phases,
and their participants.

It's not the most value-adding scope (compared to taking notes on a
piece of paper), but it lays down the foundations for further work
(in terms of code and UI), and still does _something_.

Among the interesting things in this scope are the "pretty" URLs, the
use of [Turbo Frames](https://turbo.hotwired.dev/handbook/introduction#turbo-frames%3A-decompose-complex-pages),
the use of custom RESTful actions, and the decision **not to** bother
with user accounts or character records at the moment.

(For more insights on the making of this first feature, see the commentary
track.)